pulled out methods to simply insert script
