+++ examples/setupexample-lambda/insert-example-dashboard.js
@@ -1,15 +1,8 @@
 const awsWrapper = require('./aws-wrapper');
-const crypto = require('crypto');
+const appdatabase = require('./appdatabase');
 
 const TOPIC_AREA_FILE = "topicarea.json";
 const DASHBOARD_FILE = "dashboard.json";
-
-const databaseIdPrefixes = {
-    dataset : "Dataset",
-    dashboard : "Dashboard",
-    widget : "Widget",
-    topicArea : "TopicArea",
-};
 
 class DashboardContext {
     constructor(topicAreaId, dashboardId, exampleBucketKeys) {
@@ -43,100 +36,17 @@ class DeploymentContext {
       this.tableName = tableName;
       this.createdBy = createdBy;
     }
-};
-  
-const uuidv4 = function () {
-    return crypto.randomUUID();
-}
-
-const generateDBId = function (prefix, id) {
-    
-    id = id !== undefined ? id : uuidv4();
-    return `${prefix}#${id}`;
-}
-
-const saveTopicArea = async function(deploymentContext, topicAreaBucketKey){
-
-    let topicAreaEntity = await awsWrapper.getJsonOfKey(deploymentContext.examplesBucket, topicAreaBucketKey);
-
-    topicAreaEntity.pk.S = generateDBId(databaseIdPrefixes.topicArea);
-    topicAreaEntity.sk.S = topicAreaEntity.pk.S;
-
-    topicAreaEntity.createdBy.S = deploymentContext.createdBy;
-
-    await awsWrapper.dynamoSave(deploymentContext.tableName, topicAreaEntity);
-    
-    return topicAreaEntity.pk.S;
-}
-
-const saveDashboard = async function(deploymentContext, topicAreaId, dashboardBucketKey){
-
-    let dashboardEntity = await awsWrapper.getJsonOfKey(deploymentContext.examplesBucket, dashboardBucketKey);
-
-    dashboardEntity.pk.S = generateDBId(databaseIdPrefixes.dashboard);
-    dashboardEntity.sk.S = dashboardEntity.pk.S;
-    dashboardEntity.parentDashboardId.S = dashboardEntity.pk.S;
-
-    dashboardEntity.publishedBy.S = deploymentContext.createdBy;
-    dashboardEntity.submittedBy.S = deploymentContext.createdBy;
-    dashboardEntity.createdBy.S = deploymentContext.createdBy;
-    dashboardEntity.updatedBy.S = deploymentContext.createdBy;
-
-    dashboardEntity.topicAreaId.S = topicAreaId;
-
-    await awsWrapper.dynamoSave(deploymentContext.tableName, dashboardEntity);
-    
-    return dashboardEntity.pk.S;
-}
+};  
 
 const setupDashboardTopicAreaEntities = async function(deploymentContext, exampleBucketKeys){
     
     console.log("Saving TopicArea...");
-    const topicAreaId = await saveTopicArea(deploymentContext, exampleBucketKeys.topicArea);
+    const topicAreaId = await appdatabase.saveTopicArea(deploymentContext, exampleBucketKeys.topicArea);
 
     console.log("Saving Dashboard...");
-    const dashboardId = await saveDashboard(deploymentContext, topicAreaId, exampleBucketKeys.dashboard);
+    const dashboardId = await appdatabase.saveDashboard(deploymentContext, topicAreaId, exampleBucketKeys.dashboard);
 
     return new DashboardContext(topicAreaId, dashboardId, exampleBucketKeys);
-};
-
-const saveText = async function (deploymentContext, dashboardId, widgetKey){
-
-    let widgetEntity = await awsWrapper.getJsonOfKey(deploymentContext.examplesBucket, widgetKey);
-
-    widgetEntity.pk.S = dashboardId;
-    widgetEntity.sk.S = generateDBId(databaseIdPrefixes.widget);
-
-    await awsWrapper.dynamoSave(deploymentContext.tableName, widgetEntity);
-};
-
-const saveChart = async function (deploymentContext, dashboardId, datasetUUID, datafileUUID, widgetKey){
-
-    let widgetEntity = await awsWrapper.getJsonOfKey(deploymentContext.examplesBucket, widgetKey);
-
-    widgetEntity.pk.S = dashboardId;
-    widgetEntity.sk.S = generateDBId(databaseIdPrefixes.widget);
-
-    widgetEntity.content.M.datasetId.S = datasetUUID;
-    widgetEntity.content.M.s3Key.M.raw.S = `${datafileUUID}.csv`;
-    widgetEntity.content.M.s3Key.M.json.S = `${datafileUUID}.json`;
-
-    await awsWrapper.dynamoSave(deploymentContext.tableName, widgetEntity);
-};
-
-const saveDataset = async function (deploymentContext, datasetUUID, dataFileUUID, datasetKey){
-
-    let datasetEntity = await awsWrapper.getJsonOfKey(deploymentContext.examplesBucket, datasetKey);
-
-    datasetEntity.pk.S = generateDBId(databaseIdPrefixes.dataset, datasetUUID);
-    datasetEntity.sk.S = datasetEntity.pk.S;
-    
-    datasetEntity.createdBy.S = deploymentContext.createdBy;
-
-    datasetEntity.s3Key.M.raw.S = `${dataFileUUID}.csv`;
-    datasetEntity.s3Key.M.json.S = `${dataFileUUID}.json`;
-
-    await awsWrapper.dynamoSave(deploymentContext.tableName, datasetEntity);
 };
 
 const copyContentToBucket = async function(examplesbucket, exampleFile, datasetBucket, dataFileUUID, fileExtention){
@@ -156,15 +66,15 @@ const setupExample = async function (deploymentContext, exampleBucketKeys) {
         console.log(`Setting up Widget: ${widget.name}`);
         if(widget.name.startsWith("text")){
             console.log(`Saving text for Widget: ${widget.name}`);
-            await saveText(deploymentContext, dashboardContext.dashboardId, widget.key);
+            await appdatabase.saveText(deploymentContext, dashboardContext.dashboardId, widget.key);
         }else{
-            let datasetUUID = uuidv4();
-            let dataFileUUID = uuidv4();
+            let datasetUUID = appdatabase.uuidv4();
+            let dataFileUUID = appdatabase.uuidv4();
             
             console.log(`Saving chart for Widget: ${widget.name}`);
-            await saveChart(deploymentContext, dashboardContext.dashboardId, datasetUUID, dataFileUUID, widget.key);
+            await appdatabase.saveChart(deploymentContext, dashboardContext.dashboardId, datasetUUID, dataFileUUID, widget.key);
             console.log(`Saving dataset for Widget: ${widget.name}`);
-            await saveDataset(deploymentContext, datasetUUID, dataFileUUID, widget.datasetKey);
+            await appdatabase.saveDataset(deploymentContext, datasetUUID, dataFileUUID, widget.datasetKey);
     
             let jsonDatasetKey = widget.datafiles[0].endsWith(".json") ? widget.datafiles[0] : widget.datafiles[1];
             let csvDatasetKey = widget.datafiles[0].endsWith(".csv") ? widget.datafiles[0] : widget.datafiles[1];
